+++ src/mainwindow.cpp
@@ -6777,7 +6777,14 @@ void MainWindow::on_actionExport_preview_HTML_triggered() {
  */
 void MainWindow::on_actionOpen_IRC_Channel_triggered() {
     QDesktopServices::openUrl(
-            QUrl("https://webchat.freenode.net/?channels=qownnotes"));
+            QUrl("https://kiwiirc.com/client/irc.freenode.net/#qownnotes"));
+}
+
+/**
+ * Opens the Gitter room in a webpage
+ */
+void MainWindow::on_actionGitter_triggered() {
+    QDesktopServices::openUrl(QUrl("https://gitter.im/qownnotes/qownnotes"));
 }
 
 /**